+++ app/admin-server/src/test/java/kr/co/automl/service/PreSignedUrlProviderTest.java
@@ -1,4 +1,4 @@
-package kr.co.automl.global.utils.s3;
+package kr.co.automl.service;
 
 import com.amazonaws.services.s3.AmazonS3Client;
 import org.junit.jupiter.api.BeforeEach;
